+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Fix(es):
 
 - Implement `average_time_to_start` data harvesting for the `Server` table.
 - Fix NULL/0 handling for `gpu_memory_min` and `gpu_memory_total` in vendor scripts.
+- Update normalization of benchmark scores for the workload profiles to use median ratios instead of min-max scaling.
+- Update the weighted average computation of workload profile scores to use geometric mean instead of arithmetic mean.
 
 ## v0.6.0 (June 3, 2026)